KMP 18-2 (#3734)
- FolderScreen을 여러 파일로 분리
- 폼 primitive 를 플러터 쪽과 비슷하게 함
